requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
Pillow>=8.2.0
scikit-learn>=0.24.0
faker>=8.0.0
tqdm>=4.60.0
torch>=1.9.0
transformers>=4.5.0
librosa>=0.8.0  # For audio processing
soundfile>=0.10.0  # For audio file handling
